+++ canon_keeper/repo/__init__.py
@@ -6,6 +6,7 @@ import sqlite3
 
 from canon_keeper.repo.accounts import Account, AccountRepo
 from canon_keeper.repo.campaigns import CampaignRepo
+from canon_keeper.repo.chat import ChatMessage, ChatRepo
 from canon_keeper.repo.entities import Entity, EntityRepo, StaleWrite
 from canon_keeper.repo.facts import Fact, FactRepo
 from canon_keeper.repo.layouts import LayoutRepo
@@ -22,6 +23,7 @@ class Repos:
         self.conn = conn
         self.accounts = AccountRepo(conn)
         self.campaigns = CampaignRepo(conn)
+        self.chat = ChatRepo(conn)
         self.entities = EntityRepo(conn)
         self.facts = FactRepo(conn)
         self.layouts = LayoutRepo(conn)
@@ -48,6 +50,8 @@ __all__ = [
     "UtteranceRepo",
     "Account",
     "AccountRepo",
+    "ChatMessage",
+    "ChatRepo",
     "Proposal",
     "ProposalRepo",
     "Share",